Full update of the files
Changed the GUI of the app using another library. Added a way to save the gemini response of the prompt as CLI in a .txt file.
Changed the required libraries in the requirements.txt and README.md
Added a icon to the error window when the gemini can't answer your prompt.